+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 -   Added `stripYamlFrontmatter` config option to remove YAML frontmatter from README.md, #2381.
 -   Added `--excludeCategories` config option to remove reflections present in any excluded category, #1407.
+-   If no tsconfig.json file is present, TypeDoc will now attempt to compile without setting any compiler options, #2304.
 -   Navigation is now written to a JS file and built dynamically, which significantly decreases document generation time
     with large projects and also provides large space benefits. Themes may now override `DefaultTheme.buildNavigation`
     to customize the displayed navigation tree, #2287.